fix: update example to new proc macros
